test(reporter-cvapi): move to uvu
