+++ packages/reporter-cvapi/tests/mocks/svelte.js
@@ -1,0 +1,3 @@
+import * as sinon from 'sinon';
+
+export const onDestroy = sinon.fake();
